new files of security

+++ repository/UserRepository.java
@@ -6,8 +6,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
+
+
 
     Optional<User> findByProviderIdAndProviderType(String providerId, AuthProviderType providerType);
 }
